+++ src/server/game/Server/WorldSession.h
@@ -274,7 +274,7 @@ class WorldSession
         void SendPetNameInvalid(uint32 error, std::string const& name, DeclinedName *declinedName);
         void SendPartyResult(PartyOperation operation, std::string const& member, PartyResult res, uint32 val = 0);
         void SendAreaTriggerMessage(const char* Text, ...) ATTR_PRINTF(2, 3);
-        void SendSetPhaseShift(std::set<uint32> const& phaseIds, std::set<uint32> const& terrainswaps, std::set<uint32> const& worldMapAreaSwaps);
+        void SendSetPhaseShift(std::set<uint16> const& phaseIds, std::set<uint16> const& terrainswaps, std::set<uint16> const& worldMapAreaSwaps);
         void SendQueryTimeResponse();
         void SendServerWorldInfo();
