+++ src/server/scripts/Commands/cs_debug.cpp
@@ -961,14 +961,14 @@ public:
         if (!t)
             return false;
 
-        std::set<uint32> terrainswap;
-        std::set<uint32> phaseId;
-        std::set<uint32> worldMapSwap;
+        std::set<uint16> terrainswap;
+        std::set<uint16> phaseId;
+        std::set<uint16> worldMapSwap;
 
-        terrainswap.insert((uint32)atoi(t));
+        terrainswap.insert((uint16)atoi(t));
 
         if (p)
-            phaseId.insert((uint32)atoi(p));
+            phaseId.insert((uint16)atoi(p));
 
         handler->GetSession()->SendSetPhaseShift(phaseId, terrainswap, worldMapSwap);
         return true;
